add traffic light classifier for site. trained on 1840 images
