+++ ros/src/tl_detector/launch/tl_detector.launch
@@ -1,6 +1,6 @@
 <?xml version="1.0"?>
 <launch>
     <node pkg="tl_detector" type="tl_detector.py" name="tl_detector" cwd="node">
-        <param name="inference_model" value="models/frozen_inference_graph_sim.pb" />
+        <param name="inference_class" value="TLClassifierSim" />
     </node>
 </launch>
